feat(xdr): initial commit for rust xdr serialization work
This repo is a response to the limitations found in the serde format and
rust language when handling XDR serialization. The end goal is to allow
a user to manipulate normal rust objects that can than be properly
serialized into the XDR format.

The primary hangup with serde and other implementations is the lack of
support for fixed length xdr array values. The built in rust array type
is quite limited. See:
https://stackoverflow.com/questions/30901965/implement-debug-trait-for-large-array-type
https://github.com/serde-rs/serde/issues/573

To work around this I've used Vec<T> as the type for both fixed and
variable length xdr arrays allowing a user to add a macro attribute to
distinguish between fixed/variable length while defaulting to a max
sized variable length array.

Current progress is tracked in the README.md for now, with most of the
serialization work completed and all of the deserialization work to go
still.

Breaks nothing.